+++ src/main/java/com/ibm/cloud/objectstorage/config/resource_configuration/v1/model/Firewall.java
@@ -25,8 +25,27 @@ import com.ibm.cloud.sdk.core.service.model.GenericModel;
  */
 public class Firewall extends GenericModel {
 
+  /**
+   * May contain `public`, `private`, and/or `direct` elements. Setting `allowed_network_type` to only `private` will
+   * prevent access to object storage from outside of the IBM Cloud.  The entire array will be overwritten in a `PATCH`
+   * operation. For more information on network types, [see the
+   * documentation](https://cloud.ibm.com/docs/cloud-object-storage?topic=cloud-object-storage-endpoints#advanced-endpoint-types).
+   */
+  public interface AllowedNetworkType {
+    /** public. */
+    String XPUBLIC = "public";
+    /** private. */
+    String XPRIVATE = "private";
+    /** direct. */
+    String DIRECT = "direct";
+  }
+
   @SerializedName("allowed_ip")
   private List<String> allowedIp;
+  @SerializedName("denied_ip")
+  private List<String> deniedIp;
+  @SerializedName("allowed_network_type")
+  private List<String> allowedNetworkType;
 
   /**
    * Gets the allowedIp.
@@ -41,12 +60,57 @@ public class Firewall extends GenericModel {
   }
 
   /**
+   * Gets the deniedIp.
+   *
+   * List of IPv4 or IPv6 addresses in CIDR notation to be affected by firewall in CIDR notation is supported. Passing
+   * an empty array will lift the IP address filter.  The `denied_ip` array can contain a maximum of 1000 items.
+   *
+   * @return the deniedIp
+   */
+  public List<String> getDeniedIp() {
+    return deniedIp;
+  }
+
+  /**
+   * Gets the allowedNetworkType.
+   *
+   * Indicates which network types are allowed for bucket access. May contain `public`, `private`, and/or `direct`
+   * elements. Setting `allowed_network_type` to only `private` will prevent access to object storage from outside of
+   * the IBM Cloud.  The entire array will be overwritten in a `PATCH` operation. For more information on network types,
+   * [see the
+   * documentation](https://cloud.ibm.com/docs/cloud-object-storage?topic=cloud-object-storage-endpoints#advanced-endpoint-types).
+   *
+   * @return the allowedNetworkType
+   */
+  public List<String> getAllowedNetworkType() {
+    return allowedNetworkType;
+  }
+
+  /**
    * Sets the allowedIp.
    *
    * @param allowedIp the new allowedIp
    */
   public void setAllowedIp(final List<String> allowedIp) {
     this.allowedIp = allowedIp;
+  }
+
+  /**
+   * Sets the deniedIp.
+   *
+   * @param deniedIp the new deniedIp
+   */
+  public void setDeniedIp(final List<String> deniedIp) {
+    this.deniedIp = deniedIp;
+  }
+
+  /**
+   * Sets the allowedNetworkType.
+   *
+   * @param allowedNetworkType the new allowedNetworkType
+   */
+  public void setAllowedNetworkType(final List<String> allowedNetworkType) {
+    this.allowedNetworkType = allowedNetworkType;
   }
 }